usability tweaks for build-static.sh (#2988)
* build-static.sh: -n option to only compass compile files with (n)ew changes

Default behavior is to always use --force to recompile everything.

* build-static.sh: silence errors shown if /tmp/jcompress is missing

> cat: /tmp/jcompress: No such file or directory
> bash: -c: option requires an argument

+++ bin/build-static.sh
@@ -5,9 +5,26 @@
 # 'perldoc perlartistic' or 'perldoc perlgpl'.
 
 
+force="--force"
+
+while getopts ":n" opt; do
+    case ${opt} in
+        n )
+            force=""
+            ;;
+        \? )
+            echo "$0: illegal option -- $OPTARG" 1>&2
+            exit 1
+            ;;
+    esac
+done
+
 buildroot="$LJHOME/build/static"
 mkdir -p $buildroot
-rm /tmp/jcompress
+
+if [[ -e /tmp/jcompress ]]; then
+    rm /tmp/jcompress
+fi
 
 compressor="$LJHOME/ext/yuicompressor/yuicompressor.jar"
 uncompressed_dir="/max"
@@ -26,10 +43,10 @@ if [ "$compass" != "" ]; then
     if [ $compass_version_ok ]; then
         echo "* Building SCSS..."
         cd $LJHOME
-        $compass compile -e production --force
+        $compass compile -e production $force
         if [ -d "$LJHOME/ext/dw-nonfree" ]; then
             cd $LJHOME/ext/dw-nonfree
-            $compass compile -e production --force
+            $compass compile -e production $force
         fi
     else
         echo "Compass version must be 1.0 or higher. Please upgrade."
@@ -106,8 +123,10 @@ do
     done
 
     # Now parallel execute
-    echo "Executing compression (takes a minute)..."
-    cat /tmp/jcompress | xargs -d "\n" -n 1 -P 4 -- bash -c
+    if [[ -e /tmp/jcompress ]]; then
+        echo "Executing compression (takes a minute)..."
+        cat /tmp/jcompress | xargs -d "\n" -n 1 -P 4 -- bash -c
+    fi
 done
 
 if [[ -n $compressor ]]; then
